Implement additional methods in Course controller and respectively add new methods in service and repository

+++ virtual-teacher/src/main/java/com/project/virtualteacher/entity/Course.java
@@ -41,7 +41,7 @@ public class Course {
     private boolean isPublished;
 
     @Column(name = "passing_grade")
-    private double passingGrade;
+    private int passingGrade;
 
     @Column(name = "description")
     private String description;
